@@ -8,13 +8,19 @@
     <artifactId>NexusLib</artifactId>
     <version>1.0.0</version>
     <packaging>jar</packaging>
-    
+    <!-- mvn install:install-file -Dfile=target/NexusLib-1.0.0.jar -DgroupId=tsp.nexuslib -DartifactId=NexusLib -Dversion=1.0.0 -Dpackaging=jar -->
     <name>NexusLib</name>
     <description>Plugin Library</description>
-
+    <pluginRepositories>
+        <pluginRepository>
+            <id>bytecode.space</id>
+            <url>https://repo.bytecode.space/repository/maven-public/</url>
+        </pluginRepository>
+    </pluginRepositories>
     <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
     
     <repositories>
@@ -28,19 +34,60 @@
         </repository>
     </repositories>
     
-    <dependencies>
-        <dependency>
-               <groupId>org.spigotmc</groupId>
-               <artifactId>spigot-api</artifactId>
-               <version>1.19.3-R0.1-SNAPSHOT</version>
-               <scope>provided</scope>
-        </dependency>
+    <dependencies><dependency>
+        <groupId>ca.bkaw</groupId>
+        <artifactId>paper-nms</artifactId>
+        <version>1.20.6-SNAPSHOT</version>
+        <scope>provided</scope>
+    </dependency>
         <dependency>
             <groupId>me.clip</groupId>
             <artifactId>placeholderapi</artifactId>
-            <version>2.11.3</version>
+            <version>2.11.5</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
+    <build>
+        <resources>
+            <resource>
+                <directory>src/main/resources</directory> <!-- (your resources folder location) -->
+                <filtering>true</filtering> <!-- this is the important part, it's what replaces, filters,
+                all placeholders in the resources folder (such as ${project.version} in plugin.yml) -->
+            </resource>
+        </resources>
+        <plugins>
+            <plugin>
+                <groupId>ca.bkaw</groupId>
+                <artifactId>paper-nms-maven-plugin</artifactId>
+                <version>1.4.4</version>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-source-plugin</artifactId>
+                <version>3.3.1</version>
+                <executions>
+                    <execution>
+                        <id>attach-sources</id>
+                        <goals>
+                            <goal>jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>3.4.1</version>
+                <configuration>
+                    <archive>
+                        <manifestEntries>
+                            <paperweight-mappings-namespace>mojang</paperweight-mappings-namespace>
+                            <addDefaultImplementationEntries>true</addDefaultImplementationEntries>
+                        </manifestEntries>
+                    </archive>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
