@@ -1,33 +1,38 @@
 package tsp.nexuslib.builder;
 
-import com.google.common.collect.Multimap;
-import org.bukkit.ChatColor;
-import org.bukkit.Material;
-import org.bukkit.NamespacedKey;
-import org.bukkit.OfflinePlayer;
-import org.bukkit.attribute.Attribute;
-import org.bukkit.attribute.AttributeModifier;
-import org.bukkit.enchantments.Enchantment;
-import org.bukkit.entity.Item;
-import org.bukkit.inventory.ItemFlag;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.Damageable;
-import org.bukkit.inventory.meta.ItemMeta;
-import org.bukkit.inventory.meta.SkullMeta;
-import org.bukkit.material.MaterialData;
-import org.bukkit.persistence.PersistentDataContainer;
-import org.bukkit.persistence.PersistentDataType;
-import tsp.nexuslib.util.Pair;
-import tsp.nexuslib.util.Validate;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.NamespacedKey;
+import org.bukkit.OfflinePlayer;
+import org.bukkit.attribute.Attribute;
+import org.bukkit.attribute.AttributeModifier;
+import org.bukkit.block.data.BlockData;
+import org.bukkit.enchantments.Enchantment;
+import org.bukkit.entity.Item;
+import org.bukkit.inventory.ItemFlag;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.BlockDataMeta;
+import org.bukkit.inventory.meta.Damageable;
+import org.bukkit.inventory.meta.ItemMeta;
+import org.bukkit.inventory.meta.SkullMeta;
+import org.bukkit.material.MaterialData;
+import org.bukkit.persistence.PersistentDataContainer;
+import org.bukkit.persistence.PersistentDataType;
+
+import com.google.common.collect.Multimap;
+
+import tsp.nexuslib.util.Pair;
+import tsp.nexuslib.util.Validate;
 
 /**
  * Class for building an {@link ItemStack}
@@ -45,11 +50,11 @@ public class ItemBuilder {
      *
      * @param material The material used for creating the new Object
      */
-    public ItemBuilder(@Nonnull Material material) {
+    public ItemBuilder(@Nonnull final Material material) {
         Validate.notNull(material, "Material can not be null!");
 
         this.item = new ItemStack(material);
-        this.meta = item.getItemMeta();
+        this.meta = this.item.getItemMeta();
     }
 
     /**
@@ -57,7 +62,7 @@ public class ItemBuilder {
      *
      * @param item The item used for creating the new Object
      */
-    public ItemBuilder(@Nonnull ItemStack item) {
+    public ItemBuilder(@Nonnull final ItemStack item) {
         Validate.notNull(item, "Item can not be null!");
 
         this.item = item;
@@ -65,12 +70,13 @@ public class ItemBuilder {
     }
 
     /**
-     * Creates a new ItemBuilder for the given {@link ItemStack} and {@link ItemMeta}
+     * Creates a new ItemBuilder for the given {@link ItemStack} and
+     * {@link ItemMeta}
      *
      * @param item The item used for creating the new Object
      * @param meta The ItemMeta used for creating the new Object
      */
-    public ItemBuilder(@Nonnull ItemStack item, @Nonnull ItemMeta meta) {
+    public ItemBuilder(@Nonnull final ItemStack item, @Nonnull final ItemMeta meta) {
         Validate.notNull(item, "Item can not be null!");
         Validate.notNull(meta, "Meta can not be null!");
 
@@ -83,7 +89,7 @@ public class ItemBuilder {
      *
      * @param item The item used for creating the new Object
      */
-    public ItemBuilder(@Nonnull Item item) {
+    public ItemBuilder(@Nonnull final Item item) {
         Validate.notNull(item, "Item can not be null!");
 
         this.item = item.getItemStack();
@@ -95,8 +101,8 @@ public class ItemBuilder {
      *
      * @param material The item material
      */
-    public ItemBuilder material(Material material) {
-        item.setType(material);
+    public ItemBuilder material(final Material material) {
+        this.item.withType(material);
         return this;
     }
 
@@ -105,10 +111,10 @@ public class ItemBuilder {
      *
      * @param name Name
      */
-    public ItemBuilder name(@Nonnull String name) {
+    public ItemBuilder name(@Nonnull final String name) {
         Validate.notNull(name, "Name can not be null!");
 
-        meta.setDisplayName(colorize(name));
+        this.meta.setDisplayName(this.colorize(name));
         return this;
     }
 
@@ -117,22 +123,22 @@ public class ItemBuilder {
      *
      * @param amount Amount
      */
-    public ItemBuilder amount(int amount) {
-        item.setAmount(amount);
+    public ItemBuilder amount(final int amount) {
+        this.item.setAmount(amount);
         return this;
     }
 
     /**
      * Adds an enchantment
      *
-     * @param enchantment Enchantment to add
-     * @param level Enchantment level
+     * @param enchantment            Enchantment to add
+     * @param level                  Enchantment level
      * @param ignoreLevelRestriction Should restrictions be ignored
      */
-    public ItemBuilder enchant(@Nonnull Enchantment enchantment, int level, boolean ignoreLevelRestriction) {
+    public ItemBuilder enchant(@Nonnull final Enchantment enchantment, final int level, final boolean ignoreLevelRestriction) {
         Validate.notNull(enchantment, "Enchantment can not be null!");
 
-        meta.addEnchant(enchantment, level, ignoreLevelRestriction);
+        this.meta.addEnchant(enchantment, level, ignoreLevelRestriction);
         return this;
     }
 
@@ -140,25 +146,25 @@ public class ItemBuilder {
      * Adds an enchantment
      *
      * @param enchantment Enchantment to add
-     * @param level Enchantment level
+     * @param level       Enchantment level
      */
-    public ItemBuilder enchant(@Nonnull Enchantment enchantment, int level) {
+    public ItemBuilder enchant(@Nonnull final Enchantment enchantment, final int level) {
         Validate.notNull(enchantment, "Enchantment can not be null!");
 
-        meta.addEnchant(enchantment, level, false);
+        this.meta.addEnchant(enchantment, level, false);
         return this;
     }
 
     /**
      * Adds an enchantment
      *
-     * @param enchantment Enchantment to add
+     * @param enchantment        Enchantment to add
      * @param ignoreRestrictions Should restrictions be ignored
      */
-    public ItemBuilder enchant(@Nonnull Pair<Enchantment, Integer> enchantment, boolean ignoreRestrictions) {
+    public ItemBuilder enchant(@Nonnull final Pair<Enchantment, Integer> enchantment, final boolean ignoreRestrictions) {
         Validate.notNull(enchantment, "Enchantment pair can not be null!");
 
-        meta.addEnchant(enchantment.key(), enchantment.value(), ignoreRestrictions);
+        this.meta.addEnchant(enchantment.key(), enchantment.value(), ignoreRestrictions);
         return this;
     }
 
@@ -167,10 +173,10 @@ public class ItemBuilder {
      *
      * @param enchantment Enchantment to add
      */
-    public ItemBuilder enchant(@Nonnull Pair<Enchantment, Integer> enchantment) {
+    public ItemBuilder enchant(@Nonnull final Pair<Enchantment, Integer> enchantment) {
         Validate.notNull(enchantment, "Enchantment pair can not be null!");
 
-        return enchant(enchantment.key(), enchantment.value(), false);
+        return this.enchant(enchantment.key(), enchantment.value(), false);
     }
 
     /**
@@ -178,11 +184,11 @@ public class ItemBuilder {
      *
      * @param enchantments The enchantments to add
      */
-    public ItemBuilder enchant(@Nonnull Map<Enchantment, Integer> enchantments, boolean ignoreLevelRestriction) {
+    public ItemBuilder enchant(@Nonnull final Map<Enchantment, Integer> enchantments, final boolean ignoreLevelRestriction) {
         Validate.notNull(enchantments, "Enchantments can not be null!");
 
-        for (Map.Entry<Enchantment, Integer> enchantment : enchantments.entrySet()) {
-            meta.addEnchant(enchantment.getKey(), enchantment.getValue(), ignoreLevelRestriction);
+        for (final Map.Entry<Enchantment, Integer> enchantment : enchantments.entrySet()) {
+            this.meta.addEnchant(enchantment.getKey(), enchantment.getValue(), ignoreLevelRestriction);
         }
         return this;
     }
@@ -192,10 +198,10 @@ public class ItemBuilder {
      *
      * @param enchantment The enchantment to remove
      */
-    public ItemBuilder disenchant(@Nonnull Enchantment enchantment) {
+    public ItemBuilder disenchant(@Nonnull final Enchantment enchantment) {
         Validate.notNull(enchantment, "Enchantment can not be null!");
 
-        meta.removeEnchant(enchantment);
+        this.meta.removeEnchant(enchantment);
         return this;
     }
 
@@ -204,11 +210,11 @@ public class ItemBuilder {
      *
      * @param enchantments Collection of enchantments to remove
      */
-    public ItemBuilder disenchant(@Nonnull Collection<Enchantment> enchantments) {
+    public ItemBuilder disenchant(@Nonnull final Collection<Enchantment> enchantments) {
         Validate.notNull(enchantments, "Enchantments can not be null!");
 
-        for (Enchantment enchantment : enchantments) {
-            meta.removeEnchant(enchantment);
+        for (final Enchantment enchantment : enchantments) {
+            this.meta.removeEnchant(enchantment);
         }
         return this;
     }
@@ -218,12 +224,12 @@ public class ItemBuilder {
      *
      * @param lore The string to add
      */
-    public ItemBuilder addLore(@Nonnull String lore) {
+    public ItemBuilder addLore(@Nonnull final String lore) {
         Validate.notNull(lore, "Lore can not be null!");
 
-        List<String> loreList = meta.getLore() != null ? meta.getLore() : new ArrayList<>();
-        loreList.add(colorize(lore));
-        meta.setLore(loreList);
+        final List<String> loreList = this.meta.getLore() != null ? this.meta.getLore() : new ArrayList<>();
+        loreList.add(this.colorize(lore));
+        this.meta.setLore(loreList);
         return this;
     }
 
@@ -232,15 +238,11 @@ public class ItemBuilder {
      *
      * @param lore The strings to add
      */
-    public ItemBuilder setLore(@Nullable String... lore) {
+    public ItemBuilder setLore(@Nullable final String... lore) {
         if (lore == null) {
             this.meta.setLore(null);
         } else {
-            this.meta.setLore(
-                    Arrays.stream(lore)
-                            .map(this::colorize)
-                            .collect(Collectors.toList())
-            );
+            this.meta.setLore(Arrays.stream(lore).map(this::colorize).collect(Collectors.toList()));
         }
         return this;
     }
@@ -250,15 +252,11 @@ public class ItemBuilder {
      *
      * @param lore The list to add
      */
-    public ItemBuilder setLore(@Nullable List<String> lore) {
+    public ItemBuilder setLore(@Nullable final List<String> lore) {
         if (lore == null) {
             this.meta.setLore(null);
         } else {
-            this.meta.setLore(
-                    lore.stream()
-                            .map(this::colorize)
-                            .collect(Collectors.toList())
-            );
+            this.meta.setLore(lore.stream().map(this::colorize).collect(Collectors.toList()));
         }
         return this;
     }
@@ -268,9 +266,9 @@ public class ItemBuilder {
      *
      * @param index The line to remove
      */
-    public ItemBuilder removeLore(int index) {
-        if (meta.getLore() != null) {
-            List<String> loreList = meta.getLore();
+    public ItemBuilder removeLore(final int index) {
+        if (this.meta.getLore() != null) {
+            final List<String> loreList = this.meta.getLore();
             loreList.remove(index);
             this.meta.setLore(loreList);
         }
@@ -282,11 +280,11 @@ public class ItemBuilder {
      *
      * @param line The line to remove
      */
-    public ItemBuilder removeLore(@Nonnull String line) {
+    public ItemBuilder removeLore(@Nonnull final String line) {
         Validate.notNull(line, "line can not be null!");
 
-        if (meta.getLore() != null) {
-            List<String> loreList = this.meta.getLore();
+        if (this.meta.getLore() != null) {
+            final List<String> loreList = this.meta.getLore();
             loreList.remove(line);
             this.meta.setLore(loreList);
         }
@@ -298,10 +296,10 @@ public class ItemBuilder {
      *
      * @param itemFlags The ItemFlag to add
      */
-    public ItemBuilder addItemFlags(@Nonnull ItemFlag... itemFlags) {
-        Validate.notNull(item, "ItemFlag can not be null!");
+    public ItemBuilder addItemFlags(@Nonnull final ItemFlag... itemFlags) {
+        Validate.notNull(this.item, "ItemFlag can not be null!");
 
-        meta.addItemFlags(itemFlags);
+        this.meta.addItemFlags(itemFlags);
         return this;
     }
 
@@ -310,10 +308,10 @@ public class ItemBuilder {
      *
      * @param itemFlags The ItemFlag to remove
      */
-    public ItemBuilder removeItemFlags(@Nonnull ItemFlag... itemFlags) {
+    public ItemBuilder removeItemFlags(@Nonnull final ItemFlag... itemFlags) {
         Validate.notNull(itemFlags, "ItemFlag can not be null!");
 
-        meta.removeItemFlags(itemFlags);
+        this.meta.removeItemFlags(itemFlags);
         return this;
     }
 
@@ -321,12 +319,27 @@ public class ItemBuilder {
      * Sets the material data
      *
      * @param materialData The material data to set
+     * @deprecated
      */
     @Deprecated
-    public ItemBuilder setMaterialData(@Nonnull MaterialData materialData) {
+    public ItemBuilder setMaterialData(@Nonnull final MaterialData materialData) {
         Validate.notNull(materialData, "MaterialData can not be null!");
 
-        item.setData(materialData);
+        this.item.setData(materialData);
+        return this;
+    }
+
+    /**
+     * Sets the block data
+     *
+     * @param materialData The block data to set
+     */
+    @Deprecated
+    public ItemBuilder setMaterialData(@Nonnull final BlockData materialData) {
+        Validate.notNull(materialData, "BlockData can not be null!");
+        final BlockDataMeta tmp = ((BlockDataMeta) this.item.getItemMeta());
+        tmp.setBlockData(materialData);
+        this.item.setItemMeta(tmp);
         return this;
     }
 
@@ -335,8 +348,8 @@ public class ItemBuilder {
      *
      * @param unbreakable Whether the item should be unbreakable
      */
-    public ItemBuilder setUnbreakable(boolean unbreakable) {
-        meta.setUnbreakable(unbreakable);
+    public ItemBuilder setUnbreakable(final boolean unbreakable) {
+        this.meta.setUnbreakable(unbreakable);
         return this;
     }
 
@@ -345,9 +358,10 @@ public class ItemBuilder {
      *
      * @param durability The amount of durability
      */
-    public ItemBuilder setDurability(int durability) {
-        if (!(meta instanceof Damageable)) return this;
-        ((Damageable) meta).setDamage(durability);
+    public ItemBuilder setDurability(final int durability) {
+        if (!(this.meta instanceof Damageable))
+            return this;
+        ((Damageable) this.meta).setDamage(durability);
         return this;
     }
 
@@ -355,13 +369,13 @@ public class ItemBuilder {
      * Adds an {@link Attribute}
      *
      * @param attribute The attribute to add
-     * @param modifier The attribute modifier
+     * @param modifier  The attribute modifier
      */
-    public ItemBuilder addAttributeModifier(@Nonnull Attribute attribute, @Nonnull AttributeModifier modifier) {
+    public ItemBuilder addAttributeModifier(@Nonnull final Attribute attribute, @Nonnull final AttributeModifier modifier) {
         Validate.notNull(attribute, "Attribute can not be null!");
         Validate.notNull(modifier, "Modifier can not be null!");
 
-        meta.addAttributeModifier(attribute, modifier);
+        this.meta.addAttributeModifier(attribute, modifier);
         return this;
     }
 
@@ -370,10 +384,10 @@ public class ItemBuilder {
      *
      * @param attributes The attributes to add
      */
-    public ItemBuilder setAttributeModifiers(@Nonnull Multimap<Attribute, AttributeModifier> attributes) {
+    public ItemBuilder setAttributeModifiers(@Nonnull final Multimap<Attribute, AttributeModifier> attributes) {
         Validate.notNull(attributes, "Attributes can not be null!");
 
-        meta.setAttributeModifiers(attributes);
+        this.meta.setAttributeModifiers(attributes);
         return this;
     }
 
@@ -382,11 +396,11 @@ public class ItemBuilder {
      *
      * @param glow Whether the item should glow
      */
-    public ItemBuilder setGlow(boolean glow) {
+    public ItemBuilder setGlow(final boolean glow) {
         if (glow) {
-            return enchant(item.getType() != Material.BOW ? Enchantment.ARROW_INFINITE : Enchantment.LUCK, 1, true);
+            return this.enchant(this.item.getType() != Material.BOW ? Enchantment.INFINITY : Enchantment.LUCK_OF_THE_SEA, 1, true);
         } else {
-            return disenchant(item.getType() != Material.BOW ? Enchantment.ARROW_INFINITE : Enchantment.LUCK);
+            return this.disenchant(this.item.getType() != Material.BOW ? Enchantment.INFINITY : Enchantment.LUCK_OF_THE_SEA);
         }
     }
 
@@ -395,11 +409,11 @@ public class ItemBuilder {
      *
      * @param owner The owner
      */
-    public ItemBuilder setOwner(@Nonnull OfflinePlayer owner) {
+    public ItemBuilder setOwner(@Nonnull final OfflinePlayer owner) {
         Validate.notNull(owner, "Owner can not be null!");
 
-        if (meta instanceof SkullMeta) {
-            ((SkullMeta) meta).setOwningPlayer(owner);
+        if (this.meta instanceof SkullMeta) {
+            ((SkullMeta) this.meta).setOwningPlayer(owner);
         }
         return this;
     }
@@ -409,8 +423,8 @@ public class ItemBuilder {
      *
      * @param i Model data
      */
-    public ItemBuilder setCustomModelData(int i) {
-        meta.setCustomModelData(i);
+    public ItemBuilder setCustomModelData(final int i) {
+        this.meta.setCustomModelData(i);
         return this;
     }
 
@@ -419,7 +433,7 @@ public class ItemBuilder {
      *
      * @param meta The item meta
      */
-    public ItemBuilder setItemMeta(@Nonnull ItemMeta meta) {
+    public ItemBuilder setItemMeta(@Nonnull final ItemMeta meta) {
         Validate.notNull(meta, "Meta can not be null!");
 
         this.meta = meta;
@@ -429,29 +443,29 @@ public class ItemBuilder {
     /**
      * Set persistent data
      *
-     * @param key The namespace for this data
-     * @param type The type of data
+     * @param key   The namespace for this data
+     * @param type  The type of data
      * @param value The data value
-     * @param <T> Primitive
-     * @param <Z> Complex
+     * @param <T>   Primitive
+     * @param <Z>   Complex
      */
-    public <T, Z> ItemBuilder set(NamespacedKey key, PersistentDataType<T, Z> type, Z value) {
-        getContainer().set(key, type, value);
+    public <T, Z> ItemBuilder set(final NamespacedKey key, final PersistentDataType<T, Z> type, final Z value) {
+        this.getContainer().set(key, type, value);
         return this;
     }
 
     /**
      * Set persistent data if absent
      *
-     * @param key The namespace for this data
-     * @param type The type of data
+     * @param key   The namespace for this data
+     * @param type  The type of data
      * @param value The data value
-     * @param <T> Primitive
-     * @param <Z> Complex
+     * @param <T>   Primitive
+     * @param <Z>   Complex
      */
-    public <T, Z> ItemBuilder setIfAbsent(NamespacedKey key, PersistentDataType<T, Z> type, Z value) {
-        if (!getContainer().has(key, type)) {
-            getContainer().set(key, type, value);
+    public <T, Z> ItemBuilder setIfAbsent(final NamespacedKey key, final PersistentDataType<T, Z> type, final Z value) {
+        if (!this.getContainer().has(key, type)) {
+            this.getContainer().set(key, type, value);
         }
         return this;
     }
@@ -462,8 +476,8 @@ public class ItemBuilder {
      * @return The finished item
      */
     public ItemStack build() {
-        item.setItemMeta(meta);
-        return item.clone();
+        this.item.setItemMeta(this.meta);
+        return this.item.clone();
     }
 
     /**
@@ -471,17 +485,15 @@ public class ItemBuilder {
      *
      * @return The container
      */
-    public PersistentDataContainer getContainer() {
-        return this.meta.getPersistentDataContainer();
-    }
+    public PersistentDataContainer getContainer() { return this.meta.getPersistentDataContainer(); }
 
     /**
      * Set if strings should be colorized by the builder
      *
      * @param b Whether to colorize strings
      */
-    public ItemBuilder colorize(boolean b) {
-        colorize = b;
+    public ItemBuilder colorize(final boolean b) {
+        this.colorize = b;
         return this;
     }
 
@@ -491,8 +503,9 @@ public class ItemBuilder {
      * @param string The string to colorize
      * @return Colorized string
      */
-    private String colorize(String string) {
-        if (!colorize) return string;
+    private String colorize(final String string) {
+        if (!this.colorize)
+            return string;
         return ChatColor.translateAlternateColorCodes('&', string);
     }
 
