@@ -28,63 +28,49 @@ import java.util.Optional;
 public class Config {
 
     private FileConfiguration data;
-    
-    public Config(FileConfiguration data) {
-        this.data = data;
-    }
 
-    public Config(File file) {
-        this.data = YamlConfiguration.loadConfiguration(file);
-    }
+    public Config(final FileConfiguration data) { this.data = data; }
 
-    public void reload(File file) {
-        this.data = YamlConfiguration.loadConfiguration(file);
-    }
+    public Config(final File file) { this.data = YamlConfiguration.loadConfiguration(file); }
 
-    public void reload(FileConfiguration data) {
-        this.data = data;
-    }
+    public void reload(final File file) { this.data = YamlConfiguration.loadConfiguration(file); }
+
+    public void reload(final FileConfiguration data) { this.data = data; }
 
     @Nullable
-    public Optional<Object> getOptional(@Nonnull String path) {
-        return Optional.ofNullable(data.get(path));
-    }
+    public Optional<Object> getOptional(@Nonnull final String path) { return Optional.ofNullable(this.data.get(path)); }
 
     @Nonnull
-    public <T> Optional<T> getOptionalObject(@Nonnull String path, @Nonnull Class<T> clazz) {
-        return Optional.ofNullable(data.getObject(path, clazz));
+    public <T> Optional<T> getOptionalObject(@Nonnull final String path, @Nonnull final Class<T> clazz) {
+        return Optional.ofNullable(this.data.getObject(path, clazz));
     }
 
-    public Optional<Boolean> getOptionalBoolean(@Nonnull String path) {
-        return data.contains(path) ? Optional.of(data.getBoolean(path)) : Optional.empty();
+    public Optional<Boolean> getOptionalBoolean(@Nonnull final String path) {
+        return this.data.contains(path) ? Optional.of(this.data.getBoolean(path)) : Optional.empty();
     }
 
-    public Optional<Color> getOptionalColor(@Nonnull String path) {
-        return Optional.ofNullable(data.getColor(path));
-    }
+    public Optional<Color> getOptionalColor(@Nonnull final String path) { return Optional.ofNullable(this.data.getColor(path)); }
 
-    public Optional<Double> getOptionalDouble(@Nonnull String path) {
-        if (data.get(path) instanceof Double n) {
+    public Optional<Double> getOptionalDouble(@Nonnull final String path) {
+        if (this.data.get(path) instanceof final Double n) {
             return Optional.of(n);
         } else {
             return Optional.empty();
         }
     }
 
-    public float getFloat(@Nonnull String path) {
-        return getFloat(path, 0);
-    }
+    public float getFloat(@Nonnull final String path) { return this.getFloat(path, 0); }
 
-    public float getFloat(@Nonnull String path, float def) {
-        if (data.get(path) instanceof Float n) {
+    public float getFloat(@Nonnull final String path, final float def) {
+        if (this.data.get(path) instanceof final Float n) {
             return n;
         } else {
             return def;
         }
     }
 
-    public Optional<Float> getOptionalFloat(@Nonnull String path) {
-        if (data.get(path) instanceof Float n) {
+    public Optional<Float> getOptionalFloat(@Nonnull final String path) {
+        if (this.data.get(path) instanceof final Float n) {
             return Optional.of(n);
         } else {
             return Optional.empty();
@@ -92,8 +78,8 @@ public class Config {
     }
 
     @Nonnull
-    public Optional<Integer> getOptionalInt(@Nonnull String path) {
-        if (data.get(path) instanceof Integer n) {
+    public Optional<Integer> getOptionalInt(@Nonnull final String path) {
+        if (this.data.get(path) instanceof final Integer n) {
             return Optional.of(n);
         } else {
             return Optional.empty();
@@ -101,8 +87,8 @@ public class Config {
     }
 
     @Nonnull
-    public Optional<Long> getOptionalLong(@Nonnull String path) {
-        if (data.get(path) instanceof Long n) {
+    public Optional<Long> getOptionalLong(@Nonnull final String path) {
+        if (this.data.get(path) instanceof final Long n) {
             return Optional.of(n);
         } else {
             return Optional.empty();
@@ -110,8 +96,8 @@ public class Config {
     }
 
     @Nullable
-    public Number getNumber(@Nonnull String path) {
-        if (data.get(path) instanceof Number number) {
+    public Number getNumber(@Nonnull final String path) {
+        if (this.data.get(path) instanceof final Number number) {
             return number;
         } else {
             return null;
@@ -119,20 +105,17 @@ public class Config {
     }
 
     @Nonnull
-    public Optional<Number> getOptionalNumber(@Nonnull String path) {
-        return Optional.ofNullable(getNumber(path));
-    }
+    public Optional<Number> getOptionalNumber(@Nonnull final String path) { return Optional.ofNullable(this.getNumber(path)); }
 
     @Nonnull
-    public Optional<String> getOptionalString(@Nonnull String path) {
-        return Optional.ofNullable(data.getString(path));
-    }
+    public Optional<String> getOptionalString(@Nonnull final String path) { return Optional.ofNullable(this.data.getString(path)); }
 
+    @SuppressWarnings("unchecked")
     @Nonnull
-    public <T> Optional<List<T>> getOptionalList(@Nonnull String path, Class<T> clazz) {
-        List<?> list = data.getList(path);
+    public <T> Optional<List<T>> getOptionalList(@Nonnull final String path, final Class<T> clazz) {
+        final List<?> list = this.data.getList(path);
         if (list != null && !list.isEmpty() && list.get(0).getClass().isInstance(clazz)) {
-            //noinspection unchecked
+            // noinspection unchecked
             return Optional.of((List<T>) list);
         } else {
             return Optional.empty();
@@ -140,53 +123,42 @@ public class Config {
     }
 
     @Nonnull
-    public <T extends ConfigurationSerializable> Optional<ConfigurationSerializable> getOptionalSerializable(@Nonnull String path, @Nonnull Class<T> clazz) {
-        return Optional.ofNullable(data.getSerializable(path, clazz));
+    public <T extends ConfigurationSerializable> Optional<ConfigurationSerializable> getOptionalSerializable(@Nonnull final String path,
+            @Nonnull final Class<T> clazz) {
+        return Optional.ofNullable(this.data.getSerializable(path, clazz));
     }
 
     @Nonnull
-    public Optional<Vector> getOptionalVector(@Nonnull String path) {
-        return Optional.ofNullable(data.getVector(path));
+    public Optional<Vector> getOptionalVector(@Nonnull final String path) { return Optional.ofNullable(this.data.getVector(path)); }
+
+    @Nonnull
+    public Optional<ItemStack> getOptionalItemStack(@Nonnull final String path) {
+        return Optional.ofNullable(this.data.getItemStack(path));
     }
 
     @Nonnull
-    public Optional<ItemStack> getOptionalItemStack(@Nonnull String path) {
-        return Optional.ofNullable(data.getItemStack(path));
+    public Optional<Location> getOptionalLocation(@Nonnull final String path) { return Optional.ofNullable(this.data.getLocation(path)); }
+
+    @Nonnull
+    public Optional<OfflinePlayer> getOptionalOfflinePlayer(@Nonnull final String path) {
+        return Optional.ofNullable(this.data.getOfflinePlayer(path));
     }
 
     @Nonnull
-    public Optional<Location> getOptionalLocation(@Nonnull String path) {
-        return Optional.ofNullable(data.getLocation(path));
-    }
+    public Optional<Configuration> getOptionalDefaults() { return Optional.ofNullable(this.data.getDefaults()); }
 
     @Nonnull
-    public Optional<OfflinePlayer> getOptionalOfflinePlayer(@Nonnull String path) {
-        return Optional.ofNullable(data.getOfflinePlayer(path));
-    }
+    public Optional<Configuration> getOptionalRoot() { return Optional.ofNullable(this.data.getRoot()); }
 
     @Nonnull
-    public Optional<Configuration> getOptionalDefaults() {
-        return Optional.ofNullable(data.getDefaults());
-    }
-
-    @Nonnull
-    public Optional<Configuration> getOptionalRoot() {
-        return Optional.ofNullable(data.getRoot());
-    }
-
-    @Nonnull
-    public Optional<ConfigurationSection> getOptionalConfigurationSection(@Nonnull String path) {
-        return Optional.ofNullable(data.getConfigurationSection(path));
+    public Optional<ConfigurationSection> getOptionalConfigurationSection(@Nonnull final String path) {
+        return Optional.ofNullable(this.data.getConfigurationSection(path));
     }
 
     // Convenience methods
 
-    public void setLocation(String path, Location location) {
-        data.set(path, location);
-    }
+    public void setLocation(final String path, final Location location) { this.data.set(path, location); }
 
-    public void setItemStack(String path, ItemStack itemStack) {
-        data.set(path, itemStack);
-    }
+    public void setItemStack(final String path, final ItemStack itemStack) { this.data.set(path, itemStack); }
 
 }
